// src/web3/types/types.ts
// Generated by https://quicktype.io

export interface Contract {
  name: string
  address: string
  ABI: ABI[]
  isMatic: boolean
}

export interface ABI {
  inputs: Put[]
  stateMutability?: StateMutability
  type: ABIType
  anonymous?: boolean
  name?: string
  outputs?: Put[]
}

export interface Put {
  internalType: InternalTypeEnum
  name: string
  type: InternalTypeEnum
  indexed?: boolean
}

export enum InternalTypeEnum {
  Address = 'address',
  Bool = 'bool',
  Bytes = 'bytes',
  Bytes32 = 'bytes32',
  String = 'string',
  Uint256 = 'uint256',
  Uint8 = 'uint8',
}

export enum StateMutability {
  Nonpayable = 'nonpayable',
  View = 'view',
}

export enum ABIType {
  Constructor = 'constructor',
  Event = 'event',
  Function = 'function',
}

////////////////////////////////////

// Generated by https://quicktype.io

export interface AContract {
  name: string
  address: string
  ABI: aABI[]
  isMatic: boolean
}

export interface aABI {
  inputs?: aPut[]
  stateMutability?: string
  type: string
  anonymous?: boolean
  name?: string
  outputs?: aPut[]
}

export interface aPut {
  internalType: string
  name: string
  type: string
  indexed?: boolean
}